+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## Improvements
 - #1325 Refactored CacheMachine.h and CacheMachine.cpp 
+- #1322 Updated and enabled several E2E tests
 
 ## Bug Fixes